requirements.txt
biopython==1.83
biothings-client==0.3.1
certifi==2024.6.2
charset-normalizer==3.3.2
gprofiler-official==1.0.0
idna==3.7
mygene==3.2.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
pooch==1.8.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
